moved chathistory & others to components/chat + pagination link
