+++ front/pages/w/[wId]/u/chat/[cId]/index.tsx
@@ -1,7 +1,6 @@
 import {
   ChevronDownIcon,
   ChevronRightIcon,
-  TrashIcon,
 } from "@heroicons/react/20/solid";
 import {
   ArrowRightCircleIcon,
@@ -24,12 +23,12 @@ import { Spinner } from "@app/components/Spinner";
 import TimeRangePicker, {
   ChatTimeRange,
   defaultTimeRange,
-} from "@app/components/use/ChatTimeRangePicker";
+} from "@app/components/use/chat/ChatTimeRangePicker";
 import MainTab from "@app/components/use/MainTab";
 import {
   FeedbackHandler,
   MessageFeedback,
-} from "@app/components/use/MessageFeedback";
+} from "@app/components/use/chat/MessageFeedback";
 import { runActionStreamed } from "@app/lib/actions/client";
 import {
   cloneBaseConfig,
@@ -44,18 +43,16 @@ import {
 } from "@app/lib/auth";
 import { ConnectorProvider } from "@app/lib/connectors_api";
 import { DustAPI, DustAPICredentials } from "@app/lib/dust_api";
-import { useChatSessions } from "@app/lib/swr";
 import { client_side_new_id } from "@app/lib/utils";
 import { classNames } from "@app/lib/utils";
-import { timeAgoFrom } from "@app/lib/utils";
 import {
   ChatMessageType,
   ChatRetrievedDocumentType,
-  ChatSessionType,
   MessageFeedbackStatus,
   MessageRole,
 } from "@app/types/chat";
 import { UserType, WorkspaceType } from "@app/types/user";
+import { ChatHistory } from "@app/components/use/chat/ChatHistory";
 
 const { GA_TRACKING_ID = "" } = process.env;
 
@@ -499,81 +496,6 @@ export function MessageView({
   );
 }
 
-function ChatHistory({ owner }: { owner: WorkspaceType }) {
-  const router = useRouter();
-
-  const [limit] = useState(10);
-
-  const { sessions, mutateChatSessions } = useChatSessions(owner, limit, 0);
-
-  const handleTrashClick = async (
-    event: React.MouseEvent<SVGSVGElement, MouseEvent>,
-    chatSession: ChatSessionType
-  ) => {
-    event.stopPropagation();
-    const confirmed = window.confirm(
-      `After deletion, the conversation "${chatSession.title}" cannot be recovered. Delete the conversation?`
-    );
-    if (confirmed) {
-      // call the delete API
-      const res = await fetch(
-        `/api/w/${owner.sId}/use/chats/${chatSession.sId}`,
-        {
-          method: "DELETE",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({ cId: chatSession.sId }),
-        }
-      );
-      if (res.ok) {
-        void mutateChatSessions();
-      } else {
-        const data = await res.json();
-        window.alert(`Error deleting chat: ${data.error.message}`);
-      }
-    }
-    return false;
-  };
-
-  return (
-    <div className="flex w-full flex-col">
-      {sessions && sessions.length > 0 && (
-        <>
-          <div className="mx-auto flex flex-row items-center py-8 font-bold italic">
-            Recent Chats
-          </div>
-          <div className="flex w-full flex-col space-y-2">
-            {sessions.map((s, i) => {
-              return (
-                <div
-                  key={i}
-                  className="group flex w-full cursor-pointer flex-col rounded-md border px-2 py-2 hover:bg-gray-50"
-                  onClick={() => {
-                    void router.push(`/w/${owner.sId}/u/chat/${s.sId}`);
-                  }}
-                >
-                  <div className="flex flex-row items-center">
-                    <div className="flex flex-1">{s.title}</div>
-                    <div className="min-w-16 flex flex-initial">
-                      <TrashIcon
-                        className="ml-1 hidden h-4 w-4 hover:text-violet-800 group-hover:inline-block"
-                        onClick={(e) => handleTrashClick(e, s)}
-                      ></TrashIcon>
-                      <span className="ml-2 text-xs italic text-gray-400">
-                        {timeAgoFrom(s.created)} ago
-                      </span>
-                    </div>
-                  </div>
-                </div>
-              );
-            })}
-          </div>
-        </>
-      )}
-    </div>
-  );
-}
 
 const COMMANDS: { cmd: string; description: string }[] = [
   {
